services/worker/app/requirements.txt
python-nmap
python-gvm
lxml
pika